# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi documentación de Minibares",
        "description": "Class LogController",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/inventarios": {
            "get": {
                "tags": [
                    "Invetario"
                ],
                "summary": "Mostrar información de los inventarios",
                "description": "Display a listing of the resource.",
                "operationId": "721655bcda437b9623a493da77e2fd43",
                "responses": {
                    "200": {
                        "description": "Mostrar información de los inventarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Invetario"
                ],
                "summary": "Crear un inventario",
                "description": "Store a newly created resource in storage.",
                "operationId": "e2544a669fb3cce9342d9cf42dfbf205",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "productoId": {
                                        "type": "integer"
                                    },
                                    "fechaCaducidad": {
                                        "type": "string"
                                    },
                                    "situacion": {
                                        "type": "string"
                                    },
                                    "habitacionId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "productoId": 1,
                                    "fechaCaducidad": "2020-01-01",
                                    "situacion": "A",
                                    "habitacionId": null
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crear un inventario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/inventarios/{id}": {
            "get": {
                "tags": [
                    "Invetario"
                ],
                "summary": "Mostrar información de un inventario específico",
                "description": "Display the specified resource.",
                "operationId": "61ca5382ad56e1c47831198872094a4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar información de un inventario específico."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Invetario"
                ],
                "summary": "Actualizar un inventario específico",
                "description": "Update the specified resource in storage.",
                "operationId": "dd96c5e900591f5e268aa18a8320e524",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "productoId": {
                                        "type": "integer"
                                    },
                                    "fechaCaducidad": {
                                        "type": "string"
                                    },
                                    "situacion": {
                                        "type": "string"
                                    },
                                    "habitacionId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "productoId": 1,
                                    "fechaCaducidad": "2020-01-01",
                                    "situacion": "A",
                                    "habitacionId": null
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizar un inventario específico."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Invetario"
                ],
                "summary": "Eliminar un inventario específico",
                "operationId": "72192e1c345bc6be025c87d0b8b860d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminar un inventario específico."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}